Database Re-Structuring
Update primary key definitions
==============================

Now that we have identified the target primary key definitions we can
set about implementing them as a series of Alembic migrations.

As part of handling this we have to make the following types of changes
to a number of tables:
  * add a new sequence based column to act as the primary key
  * update nullable settings for some fields, either because
    the existing settings are wrong or because they should be
    nullable, but previously were being used in primary key
    definitions and as such couldn't be configured as nullable.
  * change the primary key definition, potentially to an added
    sequence field.

This patch adds a new migration step which implements primary key
changes.

Split Servers tables into Update and Region Servers Tables
==========================================================

As part of our database re-structuring it was decided to split the
provider servers tables up into updateservers and regionservers.

This eliminates the need for the type field, and regionservers also
don't need a name field.

This patch adds a new migration step which implements the splitting
of the servers tables into update and region servers.

PINT Server Updates
===================

Re-work any code that interacted with the single servers table to
work appropriately with the split up tables. In some cases this
simplified things, e.g. no longer need to filter the single servers
table on server type to find region servers. However in other cases
this complicated matters requiring some re-structuring of the code.

Additionally the way we handle combining the shape and server_type
info to determine the appropriate type return value has evolved, as
the server type is now determined by the table the server entry comes
from, rather than a field in the table. The fact that entries that
come from a region servers table don't have name fields means that
we can leverage the existence of a name attr in combination with the
shape attr to synthesise the returned type value.

With regard to the to the /v1/<provider>/dataversion request handling
this patch aggregates update and region server versions, returning the
most recent version.
NOTE: This can be changed to instead allow for the server type to be
specified as part of the request.

Revised XML Data Importing
==========================

Fixed incremental update to work with tables whose primary key has
been converted to an integer sequence; the search for existing rows
will look for an exact match for all data fields rather that just
the primary key fields.

Updated the data_update.py XML importing to handle splitting the
servers entries up into the update and region servers as part of
the post processing.

Misc Changes
============

Removed defunct orm_exporttopostgres script.

Updated run_standalone.sh to support overriding env settings so as
to avoid the need to edit a committed file to test.

Added git ignorance of the top-level .vscode directory.

Version bumped to 2.0.4

Closes: #74, #91